+++ modules/os2forms_dawa/src/Element/DawaElementAddressMatrikula.php
@@ -151,5 +151,5 @@ class DawaElementAddressMatrikula extends WebformCompositeBase {
     }
     return $matrikula_element['matrikula'];
   }
-  
+
 }